requirements.txt
Pillow~=9.1.0
dominate~=2.6.0
imagehash~=4.2.1
